add feature to make direct access for hosts on the local network.
Note: This feature is implemented on the win32 platform only.

This feature is helpfull for the users using SOCKS to connect to host
in the firewall. If on the out side of the wall, -S option is
effective to access. If in the wall and on the same network, this
feature overrides to direct access. So we need not to modify options
in .ssh/config by the location.

Implementation: All the network interfaces are enumerated and add them
into the direct address table automatically by using GetIpAddrTable()
in IP Helper API (iphlpapi.lib). So the remote host is in the local
network, program makes direct connection whether proxy option is
specified.

Note that there is no option to avoid this behaviour currently.
I'll add it if need.