Allow cross-compiling connect
When cross-compiling e.g. on Linux for Windows, we need a way to tell
GNU make to avoid running `ver` (because that does not exist on Linux).
Likewise, the `UNAME` variable needs to be set explicitly via the
command-line if one wants to cross-compile for other platforms.

So let's make the `UNAME` and `WINVER` definitions conditional.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>